Create README.md
We have replicated the current state-of-the-art RDR prediction model and will posted it soonw. We hope that these works will promote equal comparisons of remaining mileage predictions. We refer to the original references to set hyperparameters, and these state-of-the-art models will be trained and tested under the same data conditions as the models we build